dot1x: accurate Dot1XBackend doc; de-dup constraintFor test helper
- Dot1XBackend doc comment now describes all three production backends
  (macOS cgo / EAP8021X, Windows wlanapi, other noop) instead of only darwin.
- Move the duplicated constraintFor test helper out of the darwin and windows
  test files into dot1x_helpers_test.go. It carries a `darwin || windows`
  build tag (not untagged) because it's only used by those two backends'
  tests; an untagged helper would be flagged unused on Linux.
